Updated position of cli.lua
